--- pkg/server/procedures/readme.md
@@ -1,21 +1,0 @@
-# Procedures
-
-- login
-- deploy a MachinesModel: group of machines on the same network
-  - machines could include disks, publicips, qsfss, zlogs
-  - network could have wireguard access (user access ip)
-- update MachinesModel by add a machine to a network of machines
-- update MachinesModel by remove a machine from a project
-- delete a MachinesModel by project name
-- deploy a gateway name
-- delete a gateway name by name and deployment id
-- deploy a gateway fqdn
-- delete a gateway fqdn by deployment id
-- deploy a named k8s cluster
-- update k8s by adding a node to a k8s cluster
-- update k8s by removing a node from k8s cluster
-- delete a named k8s cluster by project name
-
-- get zos deployment by deployment id
-- delete a contract
-- get all contracts
