--- pkg/server/types/gateway_fqdn.go
@@ -1,26 +1,0 @@
-package types
-
-import "github.com/threefoldtech/zos/pkg/gridtypes/zos"
-
-// GatewayFQDNModel for gateway FQDN proxy
-type GatewayFQDNModel struct {
-	// required
-	NodeID uint32 `json:"node_id"`
-	// Backends are list of backend ips
-	Backends []zos.Backend `json:"backends"`
-	// FQDN deployed on the node
-	FQDN string `json:"fqdn"`
-	// Name is the workload name
-	Name string `json:"name"`
-
-	// optional
-	// Passthrough whether to pass tls traffic or not
-	TLSPassthrough bool   `json:"tls_passthrough"`
-	Description    string `json:"description"`
-
-	// SolutionType     string
-	// NodeDeploymentID map[uint32]uint64
-
-	// computed
-	ContractID uint64 `json:"contract_id"`
-}
